RFCT: amend folder and file names for ai model in utils
